+++ nens_auth_client/models.py
@@ -207,7 +207,13 @@ class Invitation(models.Model):
         )
         return request.build_absolute_uri(relative_url)
 
-    def send_email(self, request, context=None, send_email_options=None):
+    def send_email(
+            self,
+            request,
+            context=None,
+            send_email_options=None,
+            organisation_and_roles=None,
+        ):
         """Send the invitation email.
 
         The email address is taken from invitation.email.
@@ -230,13 +236,23 @@ class Invitation(models.Model):
             the domain name from
           send_email_options (dict): an optional dict for custom send_email
             arguments. see django's docs on send_email.
+          organisation_and_roles (dict): optional dict to specify a list of
+            organisations and corresponing roles a user is invited to. 
+            format: {"org_1_name": ["role1", "role2"]}
         """
         assert self.status == self.PENDING, "The invite must be PENDING"
+
+        if organisation_and_roles:
+            organisation_and_roles = [
+                f"- {organisation}, role(s): {','.join(role)}"
+                for organisation, role in organisation_and_roles.items()
+            ]
 
         context = {
             "invitation": self,
             "accept_url": self.get_accept_url(request),
             "host": request.get_host(),
+            "organisation_and_roles": organisation_and_roles,
             **(context or {}),
         }